Strip out CC and BCC headers; otherwise staging emails will go out
Already had this fix in place, but wanted an actual fork of the original repo in version control. There is an issue on Github that's meant to address this problem, but I'm happy with a simple solution and there's appears to be more involved (and taking too long for me).